backend/requirements.txt
# --- Core FastAPI framework ---
fastapi[standard]==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1

# --- Database & ORM ---
motor==3.6.0  # async MongoDB driver

# --- Cloud storage ---
Pillow==10.4.0
boto3==1.40.39
botocore==1.40.40
google-cloud-storage==2.18.2

# --- Text-to-Speech ---
gTTS==2.5.3

# --- Redis (async caching + Upstash) ---
redis[hiredis]==5.0.8
upstash-redis==1.4.0

# --- HTTP & async helpers ---
requests==2.32.3
aiofiles==24.1.0
httpx==0.27.0

# --- Google Translator ---
# googletrans==4.0.0-rc1
# --- Translation (Google Translate via Deep Translator) ---
deep-translator==1.11.4
PyJWT==2.9.0